subsum

+++ Week-02/Day-06/subsum3.cpp
@@ -4,12 +4,12 @@ using namespace std;
 int main() {
     int n;
     cin >> n;
-    vector<int> a(n);
+    vector<long long> a(n);
     for(int i = 0; i < n; i++){
         cin >> a[i];
     }
 
-    int ans = 0;
+    long long ans = 0;
     for(int i = 0; i < n; i++){
         ans += a[i] * (i + 1) * (n - i);
     }
